refactor: Move util from LiteralExpr.as_column to LiteralValue.to_column
